:truck: move serializer to team context
