+++ api/tests/team/unit/infrastructure/serializers/jsonapi/certification-center-membership.serializer.test.js
@@ -1,6 +1,6 @@
-import { CertificationCenterMembership } from '../../../../../src/shared/domain/models/CertificationCenterMembership.js';
-import * as certificationCenterMembershipSerializer from '../../../../../src/shared/infrastructure/serializers/jsonapi/certification-center-membership.serializer.js';
-import { domainBuilder, expect } from '../../../../test-helper.js';
+import { CertificationCenterMembership } from '../../../../../../src/shared/domain/models/CertificationCenterMembership.js';
+import * as certificationCenterMembershipSerializer from '../../../../../../src/team/infrastructure/serializers/jsonapi/certification-center-membership.serializer.js';
+import { domainBuilder, expect } from '../../../../../test-helper.js';
 
 describe('Unit | Serializer | JSONAPI | certification-center-membership-serializer', function () {
   describe('#deserialize', function () {
